Правки после проверки кода
