+++ civil_registry/src/main/java/org/javaacademy/civilRegistry/model/Citizen.java
@@ -1,4 +1,4 @@
-package org.javaacademy.model;
+package org.javaacademy.civilRegistry.model;
 
 import javaacademy.human.Human;
 import lombok.Getter;
